// panel/web/src/lib/api/gen/types.gen.ts
export type Inbound = {
  id: number;
  uuid: string;
  tag: string;
  node_id: number;
  protocol: string;
  listen_port: number;
  public_port: number;
  settings: {
    [key: string]: unknown;
  };
  tls_settings: {
    [key: string]: unknown;
  };
  transport_settings: {
    [key: string]: unknown;
  };
};